fix
To reduce loop cpu checks, moveing if device_name == self.device_mac statement to the command that adds it to the list. This way the list in memory only gets commands queued up for itself since each device runs in its own thread. This will reduce the need to check quality during the loop exe action